// src/enums/events.ts
export enum Events {
  MESSAGE = 'message',
  CALLBACK_QUERY = 'callback_query',
  MESSAGE_FROM_TG = 'message-from-tg',
  VOICE_MESSAGE_FROM_TG = 'voice-message-from-tg',
  MESSAGE_TO_TG = 'message-to-tg',
  CLEAR_CHAT_HISTORY = 'clear-chat-history',
  LLM_SELECTED = 'llm-selected',
  SHOW_CURRENT_LLM = 'show-current-llm',
}